feat: Добавлены инфраструктурные зависимости и миграции
