--- app/infra/alembic/versions/2025_03_24_0236-a2ef901a576b_add_secret_key_column_into_matrix_table.py
@@ -1,47 +1,0 @@
-"""add secret key column into matrix table
-
-Revision ID: a2ef901a576b
-Revises: 35f2d8c4a31c
-Create Date: 2025-03-24 02:36:42.144073
-
-"""
-from typing import Sequence
-from typing import Union
-
-import sqlalchemy as sa
-from alembic import op
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'a2ef901a576b'
-down_revision: Union[str, None] = '35f2d8c4a31c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('user_matrix')
-    op.add_column('matrix', sa.Column('secret_key', sa.Uuid(), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('matrix', 'secret_key')
-    op.create_table(
-        'user_matrix',
-        sa.Column('uuid', sa.UUID(), autoincrement=False, nullable=False),
-        sa.Column('user_uuid', sa.UUID(), autoincrement=False, nullable=False),
-        sa.Column('matrix_uuid', sa.UUID(), autoincrement=False, nullable=False),
-        sa.ForeignKeyConstraint(
-            ['matrix_uuid'], ['matrix.uuid'], name='user_matrix_matrix_uuid_fkey'
-        ),
-        sa.ForeignKeyConstraint(
-            ['user_uuid'], ['user.uuid'], name='user_matrix_user_uuid_fkey'
-        ),
-        sa.PrimaryKeyConstraint(
-            'uuid', 'user_uuid', 'matrix_uuid', name='user_matrix_pkey'
-        ),
-    )
-    # ### end Alembic commands ###
