+++ app/infra/alembic/versions/2025_03_26_0256-e8e80ce5d350_add_secret_key_column_in_matrix.py
@@ -1,0 +1,31 @@
+"""add secret key column in matrix
+
+Revision ID: e8e80ce5d350
+Revises: 35f2d8c4a31c
+Create Date: 2025-03-26 02:56:43.320023
+
+"""
+from typing import Sequence
+from typing import Union
+
+import sqlalchemy as sa
+from alembic import op
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'e8e80ce5d350'
+down_revision: Union[str, None] = '35f2d8c4a31c'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('matrix', sa.Column('secret_key', sa.Uuid(), nullable=False))
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('matrix', 'secret_key')
+    # ### end Alembic commands ###
